+++ lib/ua_inspector/database/bots.ex
@@ -4,6 +4,7 @@ defmodule UAInspector.Database.Bots do
   """
 
   use UAInspector.Database,
+    ets_prefix: :ua_inspector_db_bots,
     sources: [{"", "bots.yml"}],
     type: :bot