+++ lib/ua_inspector/short_code_map/client_browsers.ex
@@ -4,6 +4,7 @@ defmodule UAInspector.ShortCodeMap.ClientBrowsers do
   """
 
   use UAInspector.ShortCodeMap,
+    ets_prefix: :ua_inspector_scm_client_browsers,
     file_local: "short_codes.client_browsers.yml",
     file_remote: "Parser/Client/Browser.php",
     var_name: "availableBrowsers",