+++ lib/ua_inspector/short_code_map/device_brands.ex
@@ -4,6 +4,7 @@ defmodule UAInspector.ShortCodeMap.DeviceBrands do
   """
 
   use UAInspector.ShortCodeMap,
+    ets_prefix: :ua_inspector_scm_device_brands,
     file_local: "short_codes.device_brands.yml",
     file_remote: "Parser/Device/DeviceParserAbstract.php",
     var_name: "deviceBrands",