+++ lib/ua_inspector/short_code_map/mobile_browsers.ex
@@ -4,6 +4,7 @@ defmodule UAInspector.ShortCodeMap.MobileBrowsers do
   """
 
   use UAInspector.ShortCodeMap,
+    ets_prefix: :ua_inspector_scm_mobile_browsers,
     file_local: "short_codes.mobile_browsers.yml",
     file_remote: "Parser/Client/Browser.php",
     var_name: "mobileOnlyBrowsers",