+++ lib/ua_inspector/short_code_map/oss.ex
@@ -4,6 +4,7 @@ defmodule UAInspector.ShortCodeMap.OSs do
   """
 
   use UAInspector.ShortCodeMap,
+    ets_prefix: :ua_inspector_scm_oss,
     file_local: "short_codes.oss.yml",
     file_remote: "Parser/OperatingSystem.php",
     var_name: "operatingSystems",