+++ lib/ua_inspector/storage/ets.ex
@@ -4,11 +4,25 @@ defmodule UAInspector.Storage.ETS do
   """
 
   @doc """
-  Creates a new named storage table.
+  Creates a new named storage data table.
   """
-  @spec create(module) :: :ets.tid()
-  def create(name) do
+  @spec create_data(atom) :: :ets.tid()
+  def create_data(name) do
     :ets.new(name, [:protected, :ordered_set, read_concurrency: true])
+  end
+
+  @doc """
+  Creates a new named storage lookup table.
+  """
+  @spec create_lookup(atom) :: atom
+  def create_lookup(name) do
+    case :ets.info(name) do
+      :undefined ->
+        :ets.new(name, [:named_table, :protected, :set, read_concurrency: true])
+
+      _ ->
+        name
+    end
   end
 
   @doc """
@@ -21,6 +35,33 @@ defmodule UAInspector.Storage.ETS do
         :undefined -> true
         _ -> :ets.delete(tid)
       end
+
+    :ok
+  end
+
+  @doc """
+  Tries to retrieve a data table tid from a lookup table.
+  """
+  @spec fetch_data(atom, atom) :: :ets.tid() | nil
+  def fetch_data(lookup_name, data_name) do
+    case :ets.info(lookup_name) do
+      :undefined ->
+        nil
+
+      _ ->
+        case :ets.lookup(lookup_name, data_name) do
+          [{^data_name, ets_tid}] -> ets_tid
+          _ -> nil
+        end
+    end
+  end
+
+  @doc """
+  Updates the data table tid in a lookup table.
+  """
+  @spec update_data(atom, atom, :ets.tid()) :: :ok
+  def update_data(lookup_name, data_name, data_tid) do
+    true = :ets.insert(lookup_name, {data_name, data_tid})
 
     :ok
   end